requirements.txt
python==3.11.4
streamlit==1.45.1
langchain==0.3.25
langchain-openai==0.3.17
langchain-community==0.3.24
openai==1.79.0
python-dotenv==1.1.0
wikipedia==1.4.0
duckduckgo-search==8.0.2